rename SpoilerLog.py to spoiler_log.py
(and correct all calls to it)
